fix: Changing `gallery_saver` to `gal`. #234
This was made because `gallery_saver` was "stuck" with a dependency of `http` at 0.13, not allowing to upgrade other dependencies. Check https://github.com/CarnegieTechnologies/gallery_saver/issues/205 for more information.

This commit adds `gal`, which does the same thing. `dio` and `path` were needed to recreate the behaviour of `gallery_saver`. `gal`, however, is updated and uses the latest versions of packages.